MoviesAIModel/requirements.txt
Django
djangorestframework
null
numpy
psycopg2
scikit-learn